[Bug Fix] Prevent form from submission when user press enter
Form will be refreshed if users press enter. This is not desirable because the search should be performed while users are typing.
Refer to https://github.com/erikras/redux-form/issues/1270, handleSubmit supposes to do e.preventDefault() to preview form from submitting.
Also, refer to https://github.com/erikras/redux-form/issues/190, we need to pass onSubmit as props to the form